solucion2.1/marker-service/requirements.txt
marker-pdf==1.10.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
